+++ Pandora/Paradigm/Inventory.hs
@@ -4,3 +4,4 @@ import Pandora.Paradigm.Inventory.Optics as Exports
 import Pandora.Paradigm.Inventory.Storage as Exports
 import Pandora.Paradigm.Inventory.State as Exports
 import Pandora.Paradigm.Inventory.Environment as Exports
+import Pandora.Paradigm.Inventory.Accumulator as Exports